covoar: Fix issues spotted by clang on FreeBSD 10

+++ covoar/covoar.cc
@@ -355,7 +355,9 @@ int main(
   SymbolsToAnalyze->load( symbolsFile );
   if (Verbose)
     fprintf(
-      stderr, "Analyzing %u symbols\n", SymbolsToAnalyze->set.size()
+      stderr,
+      "Analyzing %u symbols\n",
+      (unsigned int) SymbolsToAnalyze->set.size()
     );
 
   // Create explanations.
